refactor: 코드 파일 분리 및 주석 추가
