+++ AIProject/AIProject/Features/MyPage/View/ThemeView.swift
@@ -7,32 +7,7 @@
 
 import SwiftUI
 
-class ThemeManager: ObservableObject {
-    @Published var selectedTheme: Theme = .basic
-}
-
-enum Theme: String, CaseIterable {
-    case basic
-    case pop
-    case classic
-    
-    var positiveColor: Color {
-        switch self {
-        case .basic: return .aiCoPositive
-        case .pop: return .red
-        case .classic: return Color(UIColor.systemPink)
-        }
-    }
-    
-    var negativeColor: Color {
-        switch self {
-        case .basic: return .aiCoNegative
-        case .pop: return Color(UIColor.systemMint)
-        case .classic: return .green
-        }
-    }
-}
-
+/// 사용자가 테마를 선택할 수 있는 설정 뷰
 struct ThemeView: View {
     @EnvironmentObject var themeManager: ThemeManager
     
@@ -75,41 +50,4 @@ struct ThemeView: View {
 
 #Preview {
     ThemeView()
-}
-
-struct ThemeRow: View {
-    let title: String
-    let theme: Theme
-    let isSelected: Bool
-    let onTap: () -> Void
-    
-    var body: some View {
-        HStack {
-            Text(title)
-                .font(.body)
-            
-            Spacer()
-            
-            HStack(spacing: 0) {
-                theme.positiveColor
-                theme.negativeColor
-            }
-            .frame(width: 40, height: 20)
-            .clipShape(RoundedRectangle(cornerRadius: 4))
-        }
-        .padding(20)
-        .background {
-            RoundedRectangle(cornerRadius: 10)
-                .fill(isSelected ? Color.blue.opacity(0.1) : .aiCoBackground)
-        }
-        .overlay {
-            RoundedRectangle(cornerRadius: 10)
-                .stroke(isSelected ? Color.blue.opacity(0.1) : .aiCoBackground, lineWidth: 1)
-        }
-        .onTapGesture {
-            onTap()
-        }
-        .padding(.horizontal, 16)
-        .padding(.bottom, 16)
-    }
 }
